Hotfix #1
- El campo "activo" en clientes y juegos ahora se muestra correctamente
- Cambiado el diseño de todas las ventanas
- Invertida lista de cartas y la suma de las cartas en Blackjack
- Arreglado bug donde no podías poner decimales en el campo dinero/saldo

+++ src/logica/Controlador.java
@@ -126,7 +126,7 @@ public class Controlador {
 		}
 		
 		try {
-			saldo = Integer.parseInt(texto);
+			saldo = Double.parseDouble(texto);
 		
 		} catch (NumberFormatException ee) {
 			mensajeError.setText("Ingresa únicamente números");
